refactor: move focus logic to hook
